+++ index.js
@@ -1,8 +1,11 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
-import { Image, Platform, StyleSheet, TouchableOpacity, View, ViewPropTypes, NativeModules, ImageBackground } from 'react-native';
+import { Image, ImageBackground, Platform, StyleSheet, TouchableOpacity, View, ViewPropTypes } from 'react-native';
 import Icon from 'react-native-vector-icons/MaterialIcons';
-import Video from 'react-native-video';
+import Video from 'react-native-video'; // eslint-disable-line
+
+const BackgroundImage = ImageBackground || Image; // fall back to Image if RN < 0.46
+
 const styles = StyleSheet.create({
   preloadingPlaceholder: {
     backgroundColor: 'black',
@@ -138,10 +141,15 @@ export default class VideoPlayer extends Component {
   }
 
   onStartPress() {
-    this.setState({
+    if (this.props.onStart) {
+      this.props.onStart();
+    }
+
+    this.setState(state => ({
       isPlaying: true,
       isStarted: true,
-    });
+      progress: state.progress === 1 ? 0 : state.progress,
+    }));
 
     this.hideControls();
   }
@@ -169,13 +177,18 @@ export default class VideoPlayer extends Component {
 
     if (this.props.endWithThumbnail) {
       this.setState({ isStarted: false });
+      this.player.dismissFullscreenPlayer();
     }
 
-    this.player.seek(0);
+    this.setState({ progress: 1 });
+
     if (!this.props.loop) {
-      this.setState({
-        isPlaying: false,
-      });
+      this.setState(
+        { isPlaying: false },
+        () => this.player && this.player.seek(0)
+      );
+    } else {
+      this.player.seek(0);
     }
   }
 
@@ -189,6 +202,10 @@ export default class VideoPlayer extends Component {
   }
 
   onPlayPress() {
+    if (this.props.onPlayPress) {
+      this.props.onPlayPress();
+    }
+
     this.setState({
       isPlaying: !this.state.isPlaying,
     });
@@ -196,8 +213,12 @@ export default class VideoPlayer extends Component {
   }
 
   onMutePress() {
+    const isMuted = !this.state.isMuted;
+    if (this.props.onMutePress) {
+      this.props.onMutePress(isMuted);
+    }
     this.setState({
-      isMuted: !this.state.isMuted,
+      isMuted,
     });
     this.showControls();
   }
@@ -205,7 +226,7 @@ export default class VideoPlayer extends Component {
   onToggleFullScreen() {
     if(Platform.OS === "android")
     {
-      var uri = this.props.video.uri + ".mp4";
+      var uri = this.props.video.uri;
       var position = Math.floor(this.state.duration * this.state.progress);
       this.showFullscreenAndroid(uri, position);
     }
@@ -296,6 +317,10 @@ export default class VideoPlayer extends Component {
   }
 
   hideControls() {
+    if (this.props.onHideControls) {
+      this.props.onHideControls();
+    }
+
     if (this.props.disableControlsAutoHide) {
       return;
     }
@@ -310,10 +335,41 @@ export default class VideoPlayer extends Component {
   }
 
   showControls() {
+    if (this.props.onShowControls) {
+      this.props.onShowControls();
+    }
+
     this.setState({
       isControlsVisible: true,
     });
     this.hideControls();
+  }
+
+  seek(t) {
+    this.player.seek(t);
+  }
+
+  stop() {
+    this.setState({
+      isPlaying: false,
+      progress: 0,
+    });
+    this.seek(0);
+    this.showControls();
+  }
+
+  pause() {
+    this.setState({
+      isPlaying: false,
+    });
+    this.showControls();
+  }
+
+  resume() {
+    this.setState({
+      isPlaying: true,
+    });
+    this.showControls();
   }
 
   renderStartButton() {
@@ -331,7 +387,7 @@ export default class VideoPlayer extends Component {
   renderThumbnail() {
     const { thumbnail, style, customStyles, ...props } = this.props;
     return (
-      <ImageBackground
+      <BackgroundImage
         {...props}
         style={[
           styles.thumbnail,
@@ -342,7 +398,7 @@ export default class VideoPlayer extends Component {
         source={thumbnail}
       >
         {this.renderStartButton()}
-      </ImageBackground>
+      </BackgroundImage>
     );
   }
 
@@ -415,13 +471,15 @@ export default class VideoPlayer extends Component {
             />
           </TouchableOpacity>
         )}
-        <TouchableOpacity onPress={this.onToggleFullScreen} style={customStyles.controlButton}>
-          <Icon
-            style={[styles.extraControl, customStyles.controlIcon]}
-            name="fullscreen"
-            size={32}
-          />
-        </TouchableOpacity>
+        {(Platform.OS === 'android' || this.props.disableFullscreen) ? null : (
+          <TouchableOpacity onPress={this.onToggleFullScreen} style={customStyles.controlButton}>
+            <Icon
+              style={[styles.extraControl, customStyles.controlIcon]}
+              name="fullscreen"
+              size={32}
+            />
+          </TouchableOpacity>
+        )}
       </View>
     );
   }
@@ -448,7 +506,9 @@ export default class VideoPlayer extends Component {
           ]}
           ref={p => { this.player = p; }}
           muted={this.props.muted || this.state.isMuted}
-          paused={!this.state.isPlaying}
+          paused={this.props.paused
+            ? this.props.paused || !this.state.isPlaying
+            : !this.state.isPlaying}
           onProgress={this.onProgress}
           onEnd={this.onEnd}
           onLoad={this.onLoad}
@@ -512,11 +572,13 @@ VideoPlayer.propTypes = {
   videoHeight: PropTypes.number,
   duration: PropTypes.number,
   autoplay: PropTypes.bool,
+  paused: PropTypes.bool,
   defaultMuted: PropTypes.bool,
   muted: PropTypes.bool,
-  style: PropTypes.style,
+  style: ViewPropTypes.style,
   controlsTimeout: PropTypes.number,
   disableControlsAutoHide: PropTypes.bool,
+  disableFullscreen: PropTypes.bool,
   loop: PropTypes.bool,
   resizeMode: Video.propTypes.resizeMode,
   hideControlsOnStart: PropTypes.bool,
@@ -525,20 +587,20 @@ VideoPlayer.propTypes = {
   pauseOnPress: PropTypes.bool,
   fullScreenOnLongPress: PropTypes.bool,
   customStyles: PropTypes.shape({
-    wrapper: PropTypes.style,
+    wrapper: ViewPropTypes.style,
     video: Video.propTypes.style,
-    videoWrapper: PropTypes.style,
-    controls: PropTypes.style,
+    videoWrapper: ViewPropTypes.style,
+    controls: ViewPropTypes.style,
     playControl: TouchableOpacity.propTypes.style,
     controlButton: TouchableOpacity.propTypes.style,
     controlIcon: Icon.propTypes.style,
     playIcon: Icon.propTypes.style,
-    seekBar: PropTypes.style,
-    seekBarFullWidth: PropTypes.style,
-    seekBarProgress: PropTypes.style,
-    seekBarKnob: PropTypes.style,
-    seekBarKnobSeeking: PropTypes.style,
-    seekBarBackground: PropTypes.style,
+    seekBar: ViewPropTypes.style,
+    seekBarFullWidth: ViewPropTypes.style,
+    seekBarProgress: ViewPropTypes.style,
+    seekBarKnob: ViewPropTypes.style,
+    seekBarKnobSeeking: ViewPropTypes.style,
+    seekBarBackground: ViewPropTypes.style,
     thumbnail: Image.propTypes.style,
     playButton: TouchableOpacity.propTypes.style,
     playArrow: Icon.propTypes.style,
@@ -546,6 +608,11 @@ VideoPlayer.propTypes = {
   onEnd: PropTypes.func,
   onProgress: PropTypes.func,
   onLoad: PropTypes.func,
+  onStart: PropTypes.func,
+  onPlayPress: PropTypes.func,
+  onHideControls: PropTypes.func,
+  onShowControls: PropTypes.func,
+  onMutePress: PropTypes.func,
 };
 
 VideoPlayer.defaultProps = {